fixed cln-rest-plugin issue

+++ c-lightning/c-lightning-entrypoint.sh
@@ -13,14 +13,13 @@ source /usr/local/bin/create-config.sh
 
 # Install plugins first if not already installed
 CLN_REST=false
-while read -r line
-do
+while read -r line; do
     # Check line by line
     if [[ "$line" == *"# cln-rest-plugin"* ]]; then
         echo "'cln-rest-plugin' already present in config"
         CLN_REST=true
     fi
-done < "/lightningd/config"
+done <"/lightningd/config"
 
 if [ "$CLN_REST" = false ]; then
     echo "Installing 'cln-rest-plugin'..."
@@ -33,8 +32,7 @@ source /usr/local/bin/wait-for-bitcoind.sh
 echo Starting c-lightning...
 lightningd --lightning-dir=/lightningd --daemon
 
-until lightning-cli --lightning-dir=/lightningd getinfo > /dev/null 2>&1
-do
+until lightning-cli --lightning-dir=/lightningd getinfo >/dev/null 2>&1; do
     sleep 1
 done
 echo "Startup complete"
@@ -45,9 +43,13 @@ if [ -f "/lightnigd/access.macaroon" ]; then
     echo "'access.macaroon': $(xxd -ps -u -c 1000 /lightningd/access.macaroon)"
 else
     # If the file doesn't exist, move it from c-lightning-REST
-    cp /lightningd/cln-plugins/c-lightning-REST-${CLN_REST_VER}/certs/access.macaroon /root/access.macaroon
-    cp /lightningd/cln-plugins/c-lightning-REST-${CLN_REST_VER}/certs/access.macaroon /lightningd/access.macaroon
-    
+    if [ -f "/lightningd/cln-plugins/c-lightning-REST-${CLN_REST_VER}/certs/access.macaroon" ]; then
+        cp /lightningd/cln-plugins/c-lightning-REST-${CLN_REST_VER}/certs/access.macaroon /root/access.macaroon
+        cp /lightningd/cln-plugins/c-lightning-REST-${CLN_REST_VER}/certs/access.macaroon /lightningd/access.macaroon
+    else
+        echo "Macaroon not found."
+    fi
+
 fi
 
 if [[ $(lightning-cli --lightning-dir=/lightningd listfunds | jq -r ".outputs" | jq "length <= 0") == "true" ]]; then
